+++ src/CarType.ts
@@ -1,0 +1,8 @@
+export enum CarType {
+    NotSet = 'NotSet',
+    Sedan = 'Sedan',
+    SUV = 'SUV',
+    Minivan = 'Minivan',
+    Van = 'Van',
+    Truck = 'Truck'
+}
